% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandR_webtool_makeBRECI_plots.R
\name{makeBRECIplots}
\alias{makeBRECIplots}
\title{Make BRECI plots}
\usage{
makeBRECIplots(
  theseTaxa = NULL,
  currentFiles = NULL,
  futureFiles = NULL,
  outputFolder = NULL,
  NSWscope = TRUE
)
}
\arguments{
\item{theseTaxa}{A character array with the names of one or more taxa}

\item{currentFiles}{Character array. Full paths to ENM output rasters for \emph{current} climate conditions}

\item{futureFiles}{Character array. Full paths to ENM output rasters for \emph{future} climate conditions}

\item{outputFolder}{Character. Path to folder into which output PNG files will be placed}

\item{NSWscope}{Logical. Should the computation be made using the standard NSW extent? Default is TRUE.}
}
\description{
Make BRECI plots for a set of taxa contracting environmental suitability between current and future climates. By default, the scope of the computation is the NSW standard extent. If input rasters are larger than the standard NSW extent, they are cropped to size.
}
\examples{
\dontrun{}
}
