% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandR_webtool_makeClimateChangeMaps.R
\name{makeClimateChangeMaps}
\alias{makeClimateChangeMaps}
\title{Make PNG Images of Current and Future ENM models}
\usage{
makeClimateChangeMaps(
  theseTaxa = NULL,
  currentRasters = NULL,
  futureRasters = NULL,
  pngOutputPath = NULL,
  nswScope = TRUE
)
}
\arguments{
\item{theseTaxa}{Character array of taxon names}

\item{currentRasters}{Character array of full path names to rasters of current climate ENMs corresponding to the order of taxon names in \emph{theseTaxa}.}

\item{futureRasters}{Character array of full path names to rasters of future climate ENMs corresponding to the order of taxon names in \emph{theseTaxa}.}

\item{pngOutputPath}{Character object giving teh path to the folder into which PNG files will be written.}

\item{nswScope}{Logical. Should the rasters be clipped to the standard NSW extent? Default is TRUE.}
}
\description{
Make PNG Images of Current and Future ENM models
}
