% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandR_webtool_makeOccPlots.R
\name{makeOccPlot}
\alias{makeOccPlot}
\title{Make occurrence plot as a PNG file}
\usage{
makeOccPlot(
  taxonNames = NULL,
  extentStr = c("auto", "NSW", "Aust"),
  baseDataPath = NULL,
  useTaxonSubfolders = FALSE,
  outputPath = NULL
)
}
\arguments{
\item{taxonNames}{Character array. One or more taxon names to be processed}

\item{extentStr}{Character. A flag indicating the bounds to which the map will be cropped. Must be one of "auto", "NSW" or "Aust" - see \emph{Details} for more information}

\item{baseDataPath}{Character. Path to the folder below which folders for each taxon will be found}

\item{useTaxonSubfolders}{Logical. Are files in sub-folders using taxonNames entries? Default is FALSE.}

\item{outputPath}{Character. Path to the folder into which output files will be written using a filename convention described in \emph{Details}}
}
\description{
Produces a PNG image of occurrence records for a selected taxon suitable for use in the R&R webtool
}
\details{
{ The function has been designed to work with a fixed structure and filename convention for input data. It is assumed that within the folder defined by the paramater \emph{baseDataPath}, there is a sub-folder named for each taxon listed in the parameter \emph{taxonNames}. Then, within each taxon sub-folder there is a file with a name structured as "genus_specificEpithet_herbariumRecords_filtered.csv" which has been produced by the functions \emph{fetchALA} and \emph{filterALA} in the package \pkg{RandR.data}.

Output files are placed in the folder defined in paramater \emph{outputPath} with a name like "genus_specificEpithet_distibution.png". For example, "Neolitsea_dealbata_distribution.png".

The spatial bounds of the output map are set by the parameter \emph{extentStr} which may have one of three values:
\itemize{
  \item \strong{auto} Bounds set to bounding box of the occurrence records with a 0.1 degree buffer
  \item \strong{NSW}  Bounds set to encompass NSW
  \item \strong{Aust} Bounds set to encompass Australia
}
}
}
\examples{
\dontrun{}

}
