% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandR_webtool_update_taxonTable.R
\name{makeTaxonTable}
\alias{makeTaxonTable}
\title{Update the webtool taxon table}
\usage{
makeTaxonTable(outputPath = NULL)
}
\arguments{
\item{outputPath}{Character. Path to a folder into which the new version of \emph{taxonTable.csv} will be written. The folder is created if it does not exist.}
}
\description{
Make a new version of the R&R webtool taxon table from the latest R&R database table \emph{taxonTable}.
}
\details{
{ The R&R database table \emph{taxonTable} is used to make a new version of the table used to 'drive' the webtool by providing:
\itemize{
\item Canonical accepted taxon names derived from the National Species List (NSL) hosted by the Atlas of Living Australia (ALA)
\item Synonyms for accepted names to allow users some flexibility in identifying the taxon of interest
\item Addtional data for display in the webtool pages such as formatted full species name, the family to which each taxon belongs (according to the NSL), URLs to PlantNET and ALA pages for the taxon
}
The R&R database table is the upstream source for the webtool version, so adding a new taxon to the webtool must start with adding a new taxon to the R&R database version of \emph{taxonTable}. The function \emph{checkTaxonName()} in the package \emph{RandR.db}, with parameter setting 'commit == TRUE', should be used to do this.

}
}
\examples{
\dontrun{makeTaxonTable("/home/folder")}

}
