% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandR_webtool_showCovars.R
\name{showModelCovars}
\alias{showModelCovars}
\title{Show R&R GDM covariates}
\usage{
showModelCovars(thisModel)
}
\arguments{
\item{thisModel}{Character. Full path the .Rd file for an R&R GDM model object}
}
\value{
Nothing
}
\description{
A function to display the current covariate information stored in a R&R GDM model object.
}
