% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandR_webtool_trimDomain.R
\name{trim_GDM_domain}
\alias{trim_GDM_domain}
\title{Trim GDM model domain}
\usage{
trim_GDM_domain(
  thisDomainFile = NULL,
  specialZoneFile = NULL,
  makePlot = FALSE
)
}
\arguments{
\item{makePlot}{Logical. Make a plot of trimmed domain file on the standard plot device. Default is FALSE.}

\item{thisDomain}{Character. Path to a geoJSON-format spatial polygon defining the confidence bounds od "domain" for an R&R GDM}

\item{thisZone}{Character. Path to an ESRI shapefile holding a polygon layer representing "special" zones to be excluded from the confidence bounds of the GDM pointed to by thisDomain}
}
\description{
Trim a GDM model domain polygon with a special zone polygon layer
}
\details{
{ This function is designed to excise one or more polygons from the GDM bounds polygon pointed to by \emph{thisDomain}. These polygons represent areas for which it is inappropriate to produce an output from the webtool, and could include:

\itemize{
\item The range of a threatened sub-taxon;
\item One or more regions of strongly differentiated populations not formally described but suggesting the need to sample carefully; or,
\item Areas of known or inferred ploidy.
}

The intention is that if a user clicks inside one of the special regions, then an explanatory or warning popup or other message medium will be displayed.

It is assumed that if there are multiple polygons to be excised, the polygon set has been merge into one multi-polygon shapefile using the companion function \link{combineZones} or in a GIS such as ArcMap or QGIS.
}
}
\examples{
\dontrun{}
}
