% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandR_webtool_update_taxonTable.R
\name{makeTaxonTable}
\alias{makeTaxonTable}
\title{Update the webtool taxon table}
\usage{
makeTaxonTable(taxonTablePath = NULL, newTaxa = NULL)
}
\arguments{
\item{outputPath}{Character. Path to a folder into which the new version of \emph{taxonTable.csv} will be written. The folder is created if it does not exist.}
}
\description{
Make a new version of the R&R webtool taxon table from the latest R&R database table \emph{taxonTable}.
}
\details{
{ The webtool taxon table referenced in \emph{taxonTablePath} is updated to included any taxa listed in \emph{newTaxa}. The functions in R-package \link{processALA} are used to provide:
\itemize{
\item Canonical accepted taxon names derived from the National Species List (NSL) hosted by the Atlas of Living Australia (ALA)
\item Synonyms for accepted names to allow users some flexibility in identifying the taxon of interest
\item Additional data for display in the webtool pages such as formatted full species name, the family to which each taxon belongs (according to the NSL), URLs to PlantNET and ALA pages for the taxon
}
}
}
\examples{
\dontrun{

# Update existing taxa:
makeTaxonTable("/home/folder/taxonTable.csv")

# Add a new taxon:
makeTaxonTable("/home/folder/taxonTable.csv", "Pimelea spicata")

# Add several taxa:
makeTaxonTable("/home/folder/taxonTable.csv", c("Pimelea spicata", "Doryphora sassafras", "Bidens pilosa"))
}



}
